# OME-TIFF: reduce buffer size used for RandomAccessInputStreams