The buffer size should be 16 for all RandomAccessInputStream instances.
This is consistent with other TIFF-based readers, and should be faster in general than a 1MB buffer.

See testing information in https://github.com/openmicroscopy/bioformats/pull/2338.  This affects OME-TIFF only, so the other formats listed there can be ignored.