@dgault 2016-07-27
I've ran the tests against a local copy of curated/ome-tiff using test-automated with 4GB memory and 4 threads. 

The results when running with and without this PR were comparable with only a very minor speed up. I ran the tests 3 times each to ensure consistent results. Although the tests weren't hugely quicker it certainly hasn't had any negative effect in terms of performance.

Without the PR -
9 minutes 38 seconds
9 minutes 21 seconds
9 minutes 21 seconds

With the PR included -
9 minutes 15 seconds
9 minutes 10 seconds
9 minutes 15 seconds