@sbesson 2016-07-28
Also tested initialization against a [multi-files sample fileset](http://downloads.openmicroscopy.org/images/OME-TIFF/2016-06/tubhiswt-4D/) representative of the original issue with this PR:

```
[sbesson@ome-c6100-1 ~]$ for i in $(seq 1 10); do /opt/hudson/workspace/BIOFORMATS-DEV-merge-build/tools/showinf -nopix tubhiswt-4D/tubhiswt_C0_TP0.ome.tif  | grep Initialization; done
Initialization took 0.947s
Initialization took 0.657s
Initialization took 0.956s
Initialization took 0.534s
Initialization took 0.515s
Initialization took 0.664s
Initialization took 0.529s
Initialization took 0.621s
Initialization took 0.694s
Initialization took 0.654s
````

and without

```
[sbesson@ome-c6100-1 ~]$ for i in $(seq 1 10); do /opt/hudson/workspace/BIOFORMATS-DEV-latest/tools/showinf -nopix tubhiswt-4D/tubhiswt_C0_TP0.ome.tif  | grep Initialization; done
Initialization took 1.169s
Initialization took 1.317s
Initialization took 1.462s
Initialization took 1.621s
Initialization took 1.762s
Initialization took 1.271s
Initialization took 1.751s
Initialization took 2.451s
Initialization took 2.039s
Initialization took 2.176s
```

Assuming the full repository job remains green, the final OMERO-level check for this PR would be to import a representative OME-TIFF collection in a server with and without this PR and check there is no regression in 1-import time or 2- data access post-import. Our 2016-06 OME-TIFF samples might be sufficient for that purpose as they contain a variety of single and multi-file OME-TIFF /cc @pwalczysko