@sbesson 2016-08-02
Tested against the merge development server (with this PR) and the latest development server (without this PR) by importing different representative OME-TIFF samples:
- [Single file](http://downloads.openmicroscopy.org/images/OME-TIFF/2016-06/bioformats-artificial/multi-channel-4D-series.ome.tif)
- [Companion fileset](http://downloads.openmicroscopy.org/images/OME-TIFF/2016-06/companion)
- [Multi-file fileset 1](http://downloads.openmicroscopy.org/images/OME-TIFF/2016-06/tubhiswt-3D)
- [Multi-file fileset 2](http://downloads.openmicroscopy.org/images/OME-TIFF/2016-06/tubhiswt-3D).

For each sample, 3 imports were started with the `ln_s` transfer option. No negative impact on the import time due to this PR was noticed. Visualizing the imported data did not reveal any negative performance impact either. Looks good to go.